src/pipelines/real_estate_dbt/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "real-estate-dagster"
version = "1.0.0"
description = "Real Estate Data Pipeline with Dagster and dbt"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dagster>=1.5.0",
    "dagster-dbt>=0.21.0",
    "dagster-postgres>=0.21.0",
    "dbt-core>=1.6.0",
    "dbt-postgres>=1.6.0",
    "click>=8.0.0",
    "pandas>=1.5.0",
    "numpy>=1.20.0",
    "scipy>=1.9.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.0.0",
    "prophet>=1.1.0",
    "haversine>=2.8.0",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver>=1.5.0",
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["real_estate_dagster*"]

[tool.dagster]
module_name = "real_estate_dagster"
code_location_name = "real_estate"
